requirements.txt
streamlit
biopython
matplotlib
